Add files via upload
With this application you can monitoring each application is open and running in real time in the computer you can add it as Windows services and use as control and evidence in case that a user opened a restringed software. This log file is create and update each some time, you can modify the the time in order to write on the file.